Added Timers Rotation
Added Timer Script which shows a timer which also checks if the cube is in the completed state. Additionally, a new rotation manager script has been implemented allowed the user to rotate the cube, once the cube is not being rotated anymore, the cube will snap back to its default position.